pyproject.toml
[project]
name = "visionai-sdk-python"
version = "0.1.0"
description = "VisionAI SDK for Python"
readme = "README.md"
authors = [
    { name = "Tony Yang", email = "tonyyang@linkervision.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "cryptography>=46.0.5",
    "PyJWT[cryptography]>=2.8.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
